Revert "feat(aws/recon): migrate cloudfront-s3-takeover to capmodel.Risk (LAB-3995)"
