+++ go.mod
@@ -73,7 +73,6 @@ require (
 	github.com/microsoftgraph/msgraph-sdk-go-core v1.4.0
 	github.com/muesli/termenv v0.16.0
 	github.com/neo4j/neo4j-go-driver/v5 v5.28.4
-	github.com/praetorian-inc/capability-sdk v0.0.0-20260603004354-3d2abaa6bd4f
 	github.com/praetorian-inc/titus v1.1.10-0.20260324203732-67e248d9ab6b
 	github.com/spf13/cobra v1.10.2
 	github.com/stretchr/testify v1.11.1
