+++ pkg/model/model.go
@@ -3,19 +3,16 @@
 // between output, types, and plugin.
 package model
 
-// aurelianModelToken is an unexported type historically used to seal the
-// AurelianModel interface. AurelianModel is now an open (empty) interface, so
-// any type satisfies it and embedding BaseAurelianModel is no longer required.
-// This token — and the BaseAurelianModel.IsAurelianModel() helper that returns
-// it — are retained only as source-compatibility markers for existing
-// embedders, not as an enforcement mechanism.
+// aurelianModelToken is an unexported type used to seal the AurelianModel interface.
+// External packages cannot construct this type, so they cannot implement
+// AurelianModel without embedding BaseAurelianModel.
 type aurelianModelToken struct{}
 
-// AurelianModel is the marker type for Aurelian output models. It is an open
-// interface so platform types such as capmodel.Risk — external generated
-// structs we cannot attach methods to — can be emitted through the module
-// pipeline directly, without an Aurelian-side wrapper.
-type AurelianModel interface{}
+// AurelianModel is a marker interface for all Aurelian output models.
+// Only types that embed BaseAurelianModel can satisfy this interface.
+type AurelianModel interface {
+	IsAurelianModel() aurelianModelToken
+}
 
 // BaseAurelianModel is embedded into output structs to satisfy the AurelianModel interface.
 type BaseAurelianModel struct{}
